+++ tfjs-core/src/tests.ts
@@ -73,6 +73,7 @@ import './ops/greater_test';
 import './ops/image_ops_test';
 import './ops/in_top_k_test';
 import './ops/less_test';
+import './ops/less_equal_test';
 import './ops/linalg_ops_test';
 import './ops/logical_ops_test';
 import './ops/loss_ops_test';
